functioncallingをとオセロをつなげるようにした
